api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-cloud-bigquery>=3.13.0
google-auth>=2.23.0
pydantic>=2.0.0
